+++ cli/src/lib.rs
@@ -90,10 +90,10 @@
 pub use structopt;
 
 use exonum::{
-    blockchain::InstanceCollection,
+    blockchain::config::GenesisConfigBuilder,
     exonum_merkledb::{Database, RocksDB},
     node::Node,
-    runtime::{rust::ServiceFactory, Runtime},
+    runtime::{rust::ServiceFactory, RuntimeInstance, WellKnownRuntime},
 };
 use exonum_supervisor::{mode::Mode as SupervisorMode, Supervisor};
 
@@ -113,7 +113,7 @@ pub mod password;
 #[derive(Debug, Default)]
 pub struct NodeBuilder {
     services: Vec<Box<dyn ServiceFactory>>,
-    external_runtimes: Vec<(u32, Box<dyn Runtime>)>,
+    external_runtimes: Vec<RuntimeInstance>,
 }
 
 impl NodeBuilder {
@@ -128,19 +128,10 @@ impl NodeBuilder {
         self
     }
 
-    fn builtin_services(supervisor_mode: SupervisorMode) -> Vec<InstanceCollection> {
-        // Supervisor service is enabled by default.
-        let supervisor_service = match supervisor_mode {
-            SupervisorMode::Simple => Supervisor::simple_config(),
-            SupervisorMode::Decentralized => Supervisor::decentralized_config(),
-        };
-        vec![Supervisor::builtin_instance(supervisor_service)]
-    }
-
     /// Adds a new Runtime to the list of available runtimes.
     ///
     /// Note that you don't have to add a Rust Runtime, since it's included by default.
-    pub fn with_external_runtime(mut self, runtime: impl Into<(u32, Box<dyn Runtime>)>) -> Self {
+    pub fn with_external_runtime(mut self, runtime: impl WellKnownRuntime) -> Self {
         self.external_runtimes.push(runtime.into());
         self
     }
@@ -151,14 +142,22 @@ impl NodeBuilder {
     pub fn run(self) -> Result<(), failure::Error> {
         let command = Command::from_args();
 
-        let runtimes = self.external_runtimes;
-
         if let StandardResult::Run(run_config) = command.execute()? {
-            let services = Self::builtin_services(
-                run_config.node_config.public_config.general.supervisor_mode,
+            // Add builtin services to genesis config.
+            let supervisor_config =
+                match run_config.node_config.public_config.general.supervisor_mode {
+                    SupervisorMode::Simple => Supervisor::simple_config(),
+                    SupervisorMode::Decentralized => Supervisor::decentralized_config(),
+                };
+            let genesis_config = GenesisConfigBuilder::with_consensus_config(
+                run_config.node_config.public_config.consensus.clone(),
             )
-            .into_iter()
-            .chain(self.services.into_iter().map(InstanceCollection::new));
+            .with_artifact(Supervisor.artifact_id())
+            .with_instance(Supervisor::builtin_instance(supervisor_config))
+            .build();
+
+            let mut services: Vec<Box<dyn ServiceFactory>> = vec![Supervisor.into()];
+            services.extend(self.services);
 
             let db_options = &run_config.node_config.private_config.database;
             let database: Arc<dyn Database> =
@@ -169,9 +168,10 @@ impl NodeBuilder {
 
             let node = Node::new(
                 database,
-                runtimes,
+                self.external_runtimes,
                 services,
                 run_config.node_config.into(),
+                genesis_config,
                 Some(config_manager),
             );