+++ exonum/src/runtime/rust/service.rs
@@ -22,6 +22,7 @@ use std::{
 };
 
 use crate::{
+    blockchain::config::InstanceInitParams,
     crypto::{Hash, PublicKey, SecretKey},
     helpers::{Height, ValidatorId},
     messages::Verified,
@@ -86,16 +87,28 @@ pub trait Service: ServiceDispatcher + Debug + 'static {
     /// [3]: ../../blockchain/struct.Schema.html#method.state_hash_aggregator
     fn state_hash(&self, _data: BlockchainData<&dyn Snapshot>) -> Vec<Hash>;
 
+    /// Performs storage operations on behalf of the service before processing any transaction
+    /// in the block.
+    ///
+    /// Any changes of the storage state will affect `state_hash`, which means this method must
+    /// act similarly on different nodes. In other words, the service should only use data available
+    /// in the provided `CallContext`.
+    ///
+    /// The order of invoking the `before_transactions` method is an implementation detail.
+    /// Effectively, this means that services must not rely on a particular ordering
+    /// of `Service::before_transactions` invocations.
+    fn before_transactions(&self, _context: CallContext<'_>) {}
+
     /// Performs storage operations on behalf of the service before committing the block.
     ///
     /// Any changes of the storage state will affect `state_hash`, which means this method must
     /// act similarly on different nodes. In other words, the service should only use data available
-    /// in the provided `BeforeCommitContext`.
+    /// in the provided `CallContext`.
     ///
-    /// The order of invoking the `before_commit` method is an implementation detail. Effectively,
-    /// this means that services must not rely on a particular ordering of `Service::before_commit`
+    /// The order of invoking the `after_transactions` method is an implementation detail. Effectively,
+    /// this means that services must not rely on a particular ordering of `Service::after_transactions`
     /// invocations.
-    fn before_commit(&self, _context: CallContext<'_>) {}
+    fn after_transactions(&self, _context: CallContext<'_>) {}
 
     /// Handles block commit event.
     ///
@@ -134,6 +147,20 @@ where
 {
     fn from(factory: T) -> Self {
         Box::new(factory) as Self
+    }
+}
+
+/// Provides default instance configuration parameters for `ServiceFactory`.
+pub trait DefaultInstance: ServiceFactory {
+    /// Default id for a service.
+    const INSTANCE_ID: InstanceId;
+    /// Default name for a service.
+    const INSTANCE_NAME: &'static str;
+
+    /// Creates default instance configuration parameters for the service.
+    fn default_instance(&self) -> InstanceInitParams {
+        self.artifact_id()
+            .into_default_instance(Self::INSTANCE_ID, Self::INSTANCE_NAME)
     }
 }