+++ exonum/src/runtime/rust/tests.rs
@@ -28,19 +28,22 @@ use std::{
 };
 
 use crate::{
-    blockchain::{Blockchain, BlockchainMut, InstanceConfig, Schema as CoreSchema},
+    blockchain::{
+        config::{GenesisConfigBuilder, InstanceInitParams},
+        Blockchain, BlockchainMut, Schema as CoreSchema,
+    },
     helpers::{generate_testnet_config, Height, ValidatorId},
     proto::schema::tests::{TestServiceInit, TestServiceTx},
     runtime::{
         error::{ErrorKind, ExecutionError},
         BlockchainData, CallInfo, Caller, Dispatcher, DispatcherError, DispatcherSchema,
         ExecutionContext, InstanceId, InstanceSpec, InstanceStatus, Mailbox, Runtime,
-        StateHashAggregator,
+        StateHashAggregator, WellKnownRuntime,
     },
 };
 
 use super::{
-    service::{Service, ServiceFactory},
+    service::{DefaultInstance, Service, ServiceFactory},
     ArtifactId, CallContext, RustRuntime,
 };
 
@@ -87,9 +90,7 @@ fn commit_block(blockchain: &mut BlockchainMut, mut fork: Fork) {
 }
 
 fn create_runtime() -> (Inspected<RustRuntime>, Arc<Mutex<Vec<RuntimeEvent>>>) {
-    let mut runtime = RustRuntime::new(mpsc::channel(1).0);
-    let service_factory = Box::new(TestServiceImpl);
-    runtime.add_service_factory(service_factory);
+    let runtime = RustRuntime::new(mpsc::channel(1).0).with_factory(TestServiceImpl);
     let event_handle = Arc::default();
     let runtime = Inspected {
         inner: runtime,
@@ -102,14 +103,18 @@ fn create_runtime() -> (Inspected<RustRuntime>, Arc<Mutex<Vec<RuntimeEvent>>>) {
 enum RuntimeEvent {
     Initialize,
     Resume,
+    BeforeTransactions(Height, InstanceId),
     DeployArtifact(ArtifactId, Vec<u8>),
     StartAdding(InstanceSpec, Vec<u8>),
     CommitService(Option<Height>, InstanceSpec),
-    BeforeCommit(Height, InstanceId),
+    AfterTransactions(Height, InstanceId),
     AfterCommit(Height),
     Shutdown,
 }
 
+/// Test runtime wrapper logging all the events (as `RuntimeEvent`) happening within it.
+/// Other than logging, it just redirects all the calls to the inner runtime.
+/// Used to test that workflow invariants are respected.
 #[derive(Debug, Clone)]
 struct Inspected<T> {
     inner: T,
@@ -194,7 +199,7 @@ impl<T: Runtime> Runtime for Inspected<T> {
         self.inner.state_hashes(snapshot)
     }
 
-    fn before_commit(
+    fn before_transactions(
         &self,
         context: ExecutionContext<'_>,
         instance_id: u32,
@@ -203,8 +208,21 @@ impl<T: Runtime> Runtime for Inspected<T> {
         self.events
             .lock()
             .unwrap()
-            .push(RuntimeEvent::BeforeCommit(height, instance_id));
-        self.inner.before_commit(context, instance_id)
+            .push(RuntimeEvent::BeforeTransactions(height, instance_id));
+        self.inner.after_transactions(context, instance_id)
+    }
+
+    fn after_transactions(
+        &self,
+        context: ExecutionContext<'_>,
+        instance_id: u32,
+    ) -> Result<(), ExecutionError> {
+        let height = CoreSchema::new(&*context.fork).height();
+        self.events
+            .lock()
+            .unwrap()
+            .push(RuntimeEvent::AfterTransactions(height, instance_id));
+        self.inner.after_transactions(context, instance_id)
     }
 
     fn after_commit(&mut self, snapshot: &dyn Snapshot, mailbox: &mut Mailbox) {
@@ -222,10 +240,8 @@ impl<T: Runtime> Runtime for Inspected<T> {
     }
 }
 
-impl Into<(u32, Box<dyn Runtime>)> for Inspected<RustRuntime> {
-    fn into(self) -> (u32, Box<dyn Runtime>) {
-        (RustRuntime::ID as u32, Box::new(self))
-    }
+impl WellKnownRuntime for Inspected<RustRuntime> {
+    const ID: u32 = RustRuntime::ID;
 }
 
 #[derive(Debug, Clone, ProtobufConvert, BinaryValue, ObjectHash)]
@@ -328,6 +344,11 @@ impl Service for TestServiceImpl {
     }
 }
 
+impl DefaultInstance for TestServiceImpl {
+    const INSTANCE_ID: u32 = SERVICE_INSTANCE_ID;
+    const INSTANCE_NAME: &'static str = SERVICE_INSTANCE_NAME;
+}
+
 /// In this test, we manually instruct the dispatcher to deploy artifacts / create services
 /// instead of using transactions. We still need to create patches using a `BlockchainMut`
 /// in order to properly emulate the blockchain workflow.
@@ -338,9 +359,10 @@ fn basic_rust_runtime() {
     let artifact: ArtifactId = TestServiceImpl.artifact_id().into();
     // Create dummy dispatcher.
     let config = generate_testnet_config(1, 0)[0].clone();
+    let genesis_config = GenesisConfigBuilder::with_consensus_config(config.consensus).build();
     let mut blockchain = Blockchain::build_for_tests()
-        .into_mut(config.consensus)
-        .with_additional_runtime(runtime)
+        .into_mut(genesis_config)
+        .with_runtime(runtime)
         .build()
         .unwrap();
 
@@ -390,8 +412,8 @@ fn basic_rust_runtime() {
     }
     commit_block(&mut blockchain, fork);
     let events = mem::replace(&mut *event_handle.lock().unwrap(), vec![]);
-    // The service is not active at the beginning of the block, so `before_commit`
-    // should not be called for it.
+    // The service is not active at the beginning of the block, so `after_transactions`
+    // and `before_transactions` should not be called for it.
     assert_eq!(
         events,
         vec![
@@ -430,7 +452,8 @@ fn basic_rust_runtime() {
     assert_eq!(
         events,
         vec![
-            RuntimeEvent::BeforeCommit(Height(2), SERVICE_INSTANCE_ID),
+            RuntimeEvent::BeforeTransactions(Height(2), SERVICE_INSTANCE_ID),
+            RuntimeEvent::AfterTransactions(Height(2), SERVICE_INSTANCE_ID),
             RuntimeEvent::AfterCommit(Height(3)),
         ]
     );
@@ -461,7 +484,8 @@ fn basic_rust_runtime() {
     assert_eq!(
         events,
         vec![
-            RuntimeEvent::BeforeCommit(Height(3), SERVICE_INSTANCE_ID),
+            RuntimeEvent::BeforeTransactions(Height(3), SERVICE_INSTANCE_ID),
+            RuntimeEvent::AfterTransactions(Height(3), SERVICE_INSTANCE_ID),
             RuntimeEvent::AfterCommit(Height(4)),
         ]
     );
@@ -472,23 +496,21 @@ fn rust_runtime_with_builtin_services() {
     let (runtime, event_handle) = create_runtime();
     let artifact: ArtifactId = TestServiceImpl.artifact_id().into();
     let config = generate_testnet_config(1, 0)[0].clone();
-    let spec = InstanceSpec {
-        artifact: artifact.clone(),
-        id: SERVICE_INSTANCE_ID,
-        name: SERVICE_INSTANCE_NAME.to_owned(),
-    };
+    let init_params = artifact
+        .clone()
+        .into_default_instance(SERVICE_INSTANCE_ID, SERVICE_INSTANCE_NAME);
     let constructor = Init {
         msg: "constructor_message".to_owned(),
     };
 
+    let genesis_config = GenesisConfigBuilder::with_consensus_config(config.consensus.clone())
+        .with_artifact(artifact.clone())
+        .with_instance(init_params.clone().with_constructor(constructor.clone()))
+        .build();
+
     let mut blockchain = Blockchain::build_for_tests()
-        .into_mut(config.consensus.clone())
-        .with_additional_runtime(runtime)
-        .with_builtin_instances(vec![InstanceConfig::new(
-            spec.clone(),
-            Some(vec![]),
-            constructor.clone().into_bytes(),
-        )])
+        .into_mut(genesis_config.clone())
+        .with_runtime(runtime)
         .build()
         .unwrap();
 
@@ -498,8 +520,11 @@ fn rust_runtime_with_builtin_services() {
         vec![
             RuntimeEvent::Initialize,
             RuntimeEvent::DeployArtifact(artifact.clone(), vec![]),
-            RuntimeEvent::StartAdding(spec.clone(), constructor.clone().into_bytes()),
-            RuntimeEvent::CommitService(None, spec.clone()),
+            RuntimeEvent::StartAdding(
+                init_params.clone().instance_spec,
+                constructor.clone().into_bytes()
+            ),
+            RuntimeEvent::CommitService(None, init_params.clone().instance_spec),
             RuntimeEvent::AfterCommit(Height(0)),
         ]
     );
@@ -510,7 +535,8 @@ fn rust_runtime_with_builtin_services() {
     assert_eq!(
         events,
         vec![
-            RuntimeEvent::BeforeCommit(Height(0), SERVICE_INSTANCE_ID),
+            RuntimeEvent::BeforeTransactions(Height(0), SERVICE_INSTANCE_ID),
+            RuntimeEvent::AfterTransactions(Height(0), SERVICE_INSTANCE_ID),
             RuntimeEvent::AfterCommit(Height(1)),
         ]
     );
@@ -519,8 +545,8 @@ fn rust_runtime_with_builtin_services() {
     let blockchain = blockchain.inner().to_owned();
     let (runtime, event_handle) = create_runtime();
     let mut blockchain = blockchain
-        .into_mut(config.consensus)
-        .with_additional_runtime(runtime)
+        .into_mut(genesis_config)
+        .with_runtime(runtime)
         .build()
         .unwrap();
 
@@ -531,7 +557,7 @@ fn rust_runtime_with_builtin_services() {
             RuntimeEvent::Initialize,
             RuntimeEvent::DeployArtifact(artifact, vec![]),
             // `Runtime::start_adding_service` is never called for the same service
-            RuntimeEvent::CommitService(Some(Height(1)), spec),
+            RuntimeEvent::CommitService(Some(Height(1)), init_params.instance_spec),
             // `Runtime::after_commit` is never called for the same block
             RuntimeEvent::Resume,
         ]
@@ -543,7 +569,8 @@ fn rust_runtime_with_builtin_services() {
     assert_eq!(
         events,
         vec![
-            RuntimeEvent::BeforeCommit(Height(1), SERVICE_INSTANCE_ID),
+            RuntimeEvent::BeforeTransactions(Height(1), SERVICE_INSTANCE_ID),
+            RuntimeEvent::AfterTransactions(Height(1), SERVICE_INSTANCE_ID),
             RuntimeEvent::AfterCommit(Height(2)),
         ]
     );
@@ -554,9 +581,10 @@ fn conflicting_service_instances() {
     let (runtime, event_handle) = create_runtime();
     let artifact: ArtifactId = TestServiceImpl.artifact_id().into();
     let config = generate_testnet_config(1, 0)[0].clone();
+    let genesis_config = GenesisConfigBuilder::with_consensus_config(config.consensus).build();
     let mut blockchain = Blockchain::build_for_tests()
-        .into_mut(config.consensus.clone())
-        .with_additional_runtime(runtime)
+        .into_mut(genesis_config)
+        .with_runtime(runtime)
         .build()
         .unwrap();
 
@@ -669,40 +697,41 @@ impl Service for DependentServiceImpl {
     }
 }
 
-fn instance_configs() -> (InstanceConfig, InstanceConfig) {
-    let main_spec = InstanceSpec {
-        id: SERVICE_INSTANCE_ID,
-        name: SERVICE_INSTANCE_NAME.to_owned(),
-        artifact: TestServiceImpl.artifact_id().into(),
-    };
-    let dependent_spec = InstanceSpec {
-        id: SERVICE_INSTANCE_ID + 1,
-        name: "dependent-service".to_owned(),
-        artifact: DependentServiceImpl.artifact_id().into(),
-    };
-    let dependent_constructor = Init {
-        msg: SERVICE_INSTANCE_NAME.to_owned(),
-    };
+impl DefaultInstance for DependentServiceImpl {
+    const INSTANCE_ID: u32 = SERVICE_INSTANCE_ID + 1;
+    const INSTANCE_NAME: &'static str = "dependent-service";
 
-    (
-        InstanceConfig::new(main_spec, None, vec![]),
-        InstanceConfig::new(dependent_spec, None, dependent_constructor.into_bytes()),
-    )
+    fn default_instance(&self) -> InstanceInitParams {
+        self.artifact_id()
+            .into_default_instance(Self::INSTANCE_ID, Self::INSTANCE_NAME)
+            .with_constructor(Init {
+                msg: SERVICE_INSTANCE_NAME.to_owned(),
+            })
+    }
 }
 
 #[test]
 fn dependent_builtin_service() {
-    let mut runtime = RustRuntime::new(mpsc::channel(1).0);
-    runtime.add_service_factory(Box::new(TestServiceImpl));
-    runtime.add_service_factory(Box::new(DependentServiceImpl));
-    let (main_service, dep_service) = instance_configs();
+    let main_service = TestServiceImpl;
+    let dep_service = DependentServiceImpl;
 
     // Create a blockchain with both main and dependent services initialized in the genesis block.
     let config = generate_testnet_config(1, 0)[0].clone();
+
+    let genesis_config = GenesisConfigBuilder::with_consensus_config(config.consensus)
+        .with_artifact(main_service.artifact_id())
+        .with_instance(main_service.default_instance())
+        .with_artifact(dep_service.artifact_id())
+        .with_instance(dep_service.default_instance())
+        .build();
+
+    let runtime = RustRuntime::new(mpsc::channel(1).0)
+        .with_factory(main_service)
+        .with_factory(dep_service);
+
     let blockchain = Blockchain::build_for_tests()
-        .into_mut(config.consensus)
-        .with_additional_runtime(runtime)
-        .with_builtin_instances(vec![main_service, dep_service])
+        .into_mut(genesis_config)
+        .with_runtime(runtime)
         .build()
         .unwrap();
 
@@ -720,18 +749,26 @@ fn dependent_builtin_service() {
 
 #[test]
 fn dependent_builtin_service_with_incorrect_order() {
-    let mut runtime = RustRuntime::new(mpsc::channel(1).0);
-    runtime.add_service_factory(Box::new(TestServiceImpl));
-    runtime.add_service_factory(Box::new(DependentServiceImpl));
-    let (main_service, dep_service) = instance_configs();
+    let main_service = TestServiceImpl;
+    let dep_service = DependentServiceImpl;
 
     let config = generate_testnet_config(1, 0)[0].clone();
+
     // Error in the service instantiation in the genesis block bubbles up.
+    let genesis_config = GenesisConfigBuilder::with_consensus_config(config.consensus)
+        .with_artifact(main_service.artifact_id())
+        .with_artifact(dep_service.artifact_id())
+        .with_instance(dep_service.default_instance()) // <-- Incorrect service ordering
+        .with_instance(main_service.default_instance())
+        .build();
+
+    let runtime = RustRuntime::new(mpsc::channel(1).0)
+        .with_factory(main_service)
+        .with_factory(dep_service);
+
     let err = Blockchain::build_for_tests()
-        .into_mut(config.consensus)
-        .with_additional_runtime(runtime)
-        .with_builtin_instances(vec![dep_service, main_service])
-        // ^-- Incorrect service ordering
+        .into_mut(genesis_config)
+        .with_runtime(runtime)
         .build()
         .unwrap_err();
     assert!(err.to_string().contains("no dependency"));
@@ -739,26 +776,28 @@ fn dependent_builtin_service_with_incorrect_order() {
 
 #[test]
 fn dependent_service_with_no_dependency() {
-    let mut runtime = RustRuntime::new(mpsc::channel(1).0);
-    runtime.add_service_factory(Box::new(TestServiceImpl));
-    runtime.add_service_factory(Box::new(DependentServiceImpl));
-    let (_, dep_service) = instance_configs();
+    let runtime = RustRuntime::new(mpsc::channel(1).0)
+        .with_factory(TestServiceImpl)
+        .with_factory(DependentServiceImpl);
 
     let config = generate_testnet_config(1, 0)[0].clone();
+    let genesis_config = GenesisConfigBuilder::with_consensus_config(config.consensus).build();
+
     let mut blockchain = Blockchain::build_for_tests()
-        .into_mut(config.consensus)
-        .with_additional_runtime(runtime)
+        .into_mut(genesis_config)
+        .with_runtime(runtime)
         .build()
         .unwrap();
 
     let fork = create_block(&blockchain);
-    Dispatcher::commit_artifact(&fork, dep_service.instance_spec.artifact.clone(), vec![]).unwrap();
+    let inst = DependentServiceImpl.default_instance();
+    Dispatcher::commit_artifact(&fork, inst.instance_spec.artifact.clone(), vec![]).unwrap();
     commit_block(&mut blockchain, fork);
 
     let mut fork = create_block(&blockchain);
     let mut ctx = ExecutionContext::new(blockchain.dispatcher(), &mut fork, Caller::Blockchain);
     let err = ctx
-        .start_adding_service(dep_service.instance_spec, dep_service.constructor)
+        .start_adding_service(inst.instance_spec, inst.constructor)
         .unwrap_err();
     assert!(err.to_string().contains("no dependency"));
 
@@ -772,31 +811,33 @@ fn dependent_service_with_no_dependency() {
 
 #[test]
 fn dependent_service_in_same_block() {
-    let mut runtime = RustRuntime::new(mpsc::channel(1).0);
-    runtime.add_service_factory(Box::new(TestServiceImpl));
-    runtime.add_service_factory(Box::new(DependentServiceImpl));
-    let (main_service, dep_service) = instance_configs();
+    let runtime = RustRuntime::new(mpsc::channel(1).0)
+        .with_factory(TestServiceImpl)
+        .with_factory(DependentServiceImpl);
 
     let config = generate_testnet_config(1, 0)[0].clone();
+    let genesis_config = GenesisConfigBuilder::with_consensus_config(config.consensus).build();
+
     let mut blockchain = Blockchain::build_for_tests()
-        .into_mut(config.consensus)
-        .with_additional_runtime(runtime)
+        .into_mut(genesis_config)
+        .with_runtime(runtime)
         .build()
         .unwrap();
 
     // Artifacts need to be deployed in a separate block due to checks in `RustRuntime`.
     let fork = create_block(&blockchain);
-    Dispatcher::commit_artifact(&fork, main_service.instance_spec.artifact.clone(), vec![])
-        .unwrap();
-    Dispatcher::commit_artifact(&fork, dep_service.instance_spec.artifact.clone(), vec![]).unwrap();
+    let main_inst = TestServiceImpl.default_instance();
+    let dep_inst = DependentServiceImpl.default_instance();
+    Dispatcher::commit_artifact(&fork, main_inst.instance_spec.artifact.clone(), vec![]).unwrap();
+    Dispatcher::commit_artifact(&fork, dep_inst.instance_spec.artifact.clone(), vec![]).unwrap();
     commit_block(&mut blockchain, fork);
 
     // Deploy both services in the same block after genesis.
     let mut fork = create_block(&blockchain);
     let mut ctx = ExecutionContext::new(blockchain.dispatcher(), &mut fork, Caller::Blockchain);
-    ctx.start_adding_service(main_service.instance_spec, main_service.constructor)
+    ctx.start_adding_service(main_inst.instance_spec, main_inst.constructor)
         .unwrap();
-    ctx.start_adding_service(dep_service.instance_spec, dep_service.constructor)
+    ctx.start_adding_service(dep_inst.instance_spec, dep_inst.constructor)
         .unwrap();
     commit_block(&mut blockchain, fork);
 
@@ -810,26 +851,33 @@ fn dependent_service_in_same_block() {
 
 #[test]
 fn dependent_service_in_successive_block() {
-    let mut runtime = RustRuntime::new(mpsc::channel(1).0);
-    runtime.add_service_factory(Box::new(TestServiceImpl));
-    runtime.add_service_factory(Box::new(DependentServiceImpl));
-    let (main_service, dep_service) = instance_configs();
+    let main_service = TestServiceImpl;
+    let dep_service = DependentServiceImpl;
 
     let config = generate_testnet_config(1, 0)[0].clone();
+    let genesis_config = GenesisConfigBuilder::with_consensus_config(config.consensus)
+        .with_artifact(main_service.artifact_id())
+        .with_instance(main_service.default_instance())
+        .build();
+
+    let runtime = RustRuntime::new(mpsc::channel(1).0)
+        .with_factory(main_service)
+        .with_factory(dep_service);
+
     let mut blockchain = Blockchain::build_for_tests()
-        .into_mut(config.consensus)
-        .with_additional_runtime(runtime)
-        .with_builtin_instances(vec![main_service])
+        .into_mut(genesis_config)
+        .with_runtime(runtime)
         .build()
         .unwrap();
 
     let fork = create_block(&blockchain);
-    Dispatcher::commit_artifact(&fork, dep_service.instance_spec.artifact.clone(), vec![]).unwrap();
+    let dep_spec = DependentServiceImpl.default_instance();
+    Dispatcher::commit_artifact(&fork, dep_spec.instance_spec.artifact.clone(), vec![]).unwrap();
     commit_block(&mut blockchain, fork);
 
     let mut fork = create_block(&blockchain);
     ExecutionContext::new(blockchain.dispatcher(), &mut fork, Caller::Blockchain)
-        .start_adding_service(dep_service.instance_spec, dep_service.constructor)
+        .start_adding_service(dep_spec.instance_spec, dep_spec.constructor)
         .unwrap();
     commit_block(&mut blockchain, fork);