+++ exonum/src/runtime/types.rs
@@ -23,7 +23,7 @@ use serde_derive::{Deserialize, Serialize};
 use std::{borrow::Cow, fmt::Display, str::FromStr};
 
 use super::InstanceDescriptor;
-use crate::{helpers::ValidateInput, proto::schema};
+use crate::{blockchain::config::InstanceInitParams, helpers::ValidateInput, proto::schema};
 
 /// Unique service instance identifier.
 ///
@@ -165,6 +165,15 @@ impl ArtifactId {
         Ok(artifact)
     }
 
+    /// Converts into `InstanceInitParams` with given id, name and empty constructor.
+    pub fn into_default_instance(
+        self,
+        id: InstanceId,
+        name: impl Into<String>,
+    ) -> InstanceInitParams {
+        InstanceInitParams::new(id, name, self, ())
+    }
+
     /// Checks that the artifact name contains only allowed characters and is not empty.
     fn is_valid_name(name: impl AsRef<[u8]>) -> bool {
         // Extended version of `exonum_merkledb::is_valid_name` that also allows '.' and ':'.
@@ -229,13 +238,34 @@ impl FromStr for ArtifactId {
 }
 
 /// Exhaustive artifact specification.
-#[derive(Debug, Clone, PartialEq, Eq, Hash, ProtobufConvert, BinaryValue, ObjectHash)]
+#[derive(
+    Debug,
+    Clone,
+    PartialEq,
+    Eq,
+    Hash,
+    Serialize,
+    Deserialize,
+    ProtobufConvert,
+    BinaryValue,
+    ObjectHash,
+)]
 #[protobuf_convert(source = "schema::runtime::ArtifactSpec")]
 pub struct ArtifactSpec {
     /// Information uniquely identifying the artifact.
     pub artifact: ArtifactId,
     /// Runtime-specific artifact payload.
     pub payload: Vec<u8>,
+}
+
+impl ArtifactSpec {
+    /// Generic constructor.
+    pub fn new(artifact: ArtifactId, deploy_spec: impl BinaryValue) -> Self {
+        Self {
+            artifact,
+            payload: deploy_spec.into_bytes(),
+        }
+    }
 }
 
 /// Exhaustive service instance specification.
@@ -357,8 +387,8 @@ pub enum ArtifactStatus {
 impl Display for ArtifactStatus {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
-            Self::Active => f.write_str("active"),
-            Self::Pending => f.write_str("pending"),
+            ArtifactStatus::Active => f.write_str("active"),
+            ArtifactStatus::Pending => f.write_str("pending"),
         }
     }
 }
@@ -368,15 +398,15 @@ impl ProtobufConvert for ArtifactStatus {
 
     fn to_pb(&self) -> Self::ProtoStruct {
         match self {
-            Self::Active => Self::ProtoStruct::ARTIFACT_ACTIVE,
-            Self::Pending => Self::ProtoStruct::ARTIFACT_PENDING,
+            ArtifactStatus::Active => schema::runtime::ArtifactStatus::ARTIFACT_ACTIVE,
+            ArtifactStatus::Pending => schema::runtime::ArtifactStatus::ARTIFACT_PENDING,
         }
     }
 
     fn from_pb(pb: Self::ProtoStruct) -> Result<Self, failure::Error> {
         Ok(match pb {
-            Self::ProtoStruct::ARTIFACT_ACTIVE => Self::Active,
-            Self::ProtoStruct::ARTIFACT_PENDING => Self::Pending,
+            schema::runtime::ArtifactStatus::ARTIFACT_ACTIVE => ArtifactStatus::Active,
+            schema::runtime::ArtifactStatus::ARTIFACT_PENDING => ArtifactStatus::Pending,
         })
     }
 }
@@ -393,8 +423,8 @@ pub enum InstanceStatus {
 impl Display for InstanceStatus {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
-            Self::Active => f.write_str("active"),
-            Self::Pending => f.write_str("pending"),
+            InstanceStatus::Active => f.write_str("active"),
+            InstanceStatus::Pending => f.write_str("pending"),
         }
     }
 }
@@ -404,15 +434,15 @@ impl ProtobufConvert for InstanceStatus {
 
     fn to_pb(&self) -> Self::ProtoStruct {
         match self {
-            Self::Active => Self::ProtoStruct::SERVICE_ACTIVE,
-            Self::Pending => Self::ProtoStruct::SERVICE_PENDING,
+            InstanceStatus::Active => schema::runtime::InstanceStatus::SERVICE_ACTIVE,
+            InstanceStatus::Pending => schema::runtime::InstanceStatus::SERVICE_PENDING,
         }
     }
 
     fn from_pb(pb: Self::ProtoStruct) -> Result<Self, failure::Error> {
         Ok(match pb {
-            Self::ProtoStruct::SERVICE_ACTIVE => Self::Active,
-            Self::ProtoStruct::SERVICE_PENDING => Self::Pending,
+            schema::runtime::InstanceStatus::SERVICE_ACTIVE => InstanceStatus::Active,
+            schema::runtime::InstanceStatus::SERVICE_PENDING => InstanceStatus::Pending,
         })
     }
 }