+++ services/supervisor/tests/supervisor/config.rs
@@ -16,10 +16,12 @@ use exonum_merkledb::ObjectHash;
 use exonum_testkit::TestKitBuilder;
 
 use exonum::{
-    blockchain::InstanceCollection,
     crypto,
     helpers::{Height, ValidatorId},
-    runtime::{rust::Transaction, InstanceId, SUPERVISOR_INSTANCE_ID},
+    runtime::{
+        rust::{ServiceFactory, Transaction},
+        InstanceId, SUPERVISOR_INSTANCE_ID,
+    },
 };
 
 use crate::{utils::*, IncService as ConfigChangeService};
@@ -463,15 +465,19 @@ fn test_another_configuration_change_proposal() {
 fn test_service_config_discard_fake_supervisor() {
     const FAKE_SUPERVISOR_ID: InstanceId = 5;
     let keypair = crypto::gen_keypair();
+    let fake_supervisor_artifact = Supervisor.artifact_id();
+
+    let fake_supervisor_instance = fake_supervisor_artifact
+        .clone()
+        .into_default_instance(FAKE_SUPERVISOR_ID, "fake-supervisor")
+        .with_constructor(Supervisor::decentralized_config());
 
     let mut testkit = TestKitBuilder::validator()
         .with_validators(1)
-        .with_rust_service(InstanceCollection::new(Supervisor).with_instance(
-            FAKE_SUPERVISOR_ID,
-            "fake-supervisor",
-            Vec::default(),
-        ))
-        .with_rust_service(ConfigChangeService)
+        .with_rust_service(Supervisor)
+        .with_artifact(fake_supervisor_artifact)
+        .with_instance(fake_supervisor_instance)
+        .with_default_rust_service(ConfigChangeService)
         .create();
 
     let params = "I am a new parameter".to_owned();