+++ CHANGELOG.md
@@ -9,6 +9,29 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 #### exonum
 
+- `before_commit` hook was renamed to the `after_transactions`. (#1577)
+
+#### exonum-supervisor
+
+- `Supervisor` structure isn't generic anymore. (#1587)
+
+### New features
+
+#### exonum
+
+- `before_transactions` hook for services was introduced. (#1577)
+
+#### exonum-supervisor
+
+- `Supervisor` service now can has initial configuration and implements
+  `Configure` interface. (#1587)
+
+## 0.13.0-rc.2 - 2019-12-04
+
+### Breaking changes
+
+#### exonum
+
 - **Most important**: new Dynamic Services feature was introduced. For details see
   the [Dynamic Services](#dynamic-services-feature) section of the changelog.
 
@@ -51,6 +74,11 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 - Services can now use `BlockchainData` and `SnapshotExt` types to access data
   from the blockchain in a more structured manner. (#1523)
+
+- `GenesisConfig` is extracted into separate entity. `BlockchainBuilder`, `Node`
+ and `Testkit` explicitly accepts it during creation. (#1541)
+
+- Added `DefaultInstance` trait for declaration of builtin services. (#1541)
 
 #### exonum-merkledb
 
@@ -123,8 +151,8 @@ Key points of this feature are the following:
   For details see the [`Runtime` trait docs][runtime-trait] and the
   [`sample_runtime` example][sample-runtime].
 
-  [runtime-trait]: https://docs.rs/exonum/0.13.0/exonum/runtime/trait.Runtime.html
-  [sample-runtime]: https://github.com/exonum/exonum/tree/v0.13/examples/sample_runtime
+  [runtime-trait]: https://docs.rs/exonum/0.13.0-rc.2/exonum/runtime/trait.Runtime.html
+  [sample-runtime]: https://github.com/exonum/exonum/tree/v0.13.0-rc.2/examples/sample_runtime
 
 - Services are not statically tied to the compiled binary anymore. There is
   support of adding new service types (aka artifacts) dynamically and starting new
@@ -132,7 +160,7 @@ Key points of this feature are the following:
 
   For details see [`runtime` module docs][runtime-docs].
 
-  [runtime-docs]: https://docs.rs/exonum/0.13.0/exonum/runtime/index.html
+  [runtime-docs]: https://docs.rs/exonum/0.13.0-rc.2/exonum/runtime/index.html
 
 - Services now can have initialization parameters, provided within service start
   procedure.
@@ -143,7 +171,7 @@ Key points of this feature are the following:
   capable of not only changing configuration, but of deploying and starting
   services as well. For details see [`supervisor` service][supervisor].
 
-  [supervisor]: https://github.com/exonum/exonum/tree/v0.13/services/supervisor
+  [supervisor]: https://github.com/exonum/exonum/tree/v0.13.0-rc.2/services/supervisor
 
 #### Migration Guide
 
@@ -152,7 +180,7 @@ So to make the changes apparent, compare the `Cryptocurrency` example service ve
 for [0.12.1][crypt-0-12] and [0.13.0][crypt-0-13] releases.
 
 [crypt-0-12]: https://github.com/exonum/exonum/blob/v0.12.1/examples/cryptocurrency/
-[crypt-0-13]: https://github.com/exonum/exonum/blob/v0.13/examples/cryptocurrency/
+[crypt-0-13]: https://github.com/exonum/exonum/blob/v0.13.0-rc.2/examples/cryptocurrency/
 
 Key points:
 
@@ -458,6 +486,8 @@ Key points:
 - #1555: Update MerkleDB docs
 
 - #1568: Make DispatcherSchema merkelized again [ECR-3810]
+
+- #1592: Fix node freeze after re-enabling consensus [ERC-3111]
 
 </details>
 <!-- markdownlint-enable no-inline-html -->